server/requirements.txt
Django>=5.1
django-ninja>=1.3
django-allauth[google]>=65.0
django-cors-headers>=4.4
PyJWT>=2.9
python-dotenv>=1.0
Pillow>=10.0
boto3>=1.35
